requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pandas==2.2.0
openpyxl==3.1.2
gitpython==3.1.41
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1
cryptography==42.0.5
python-docx==1.1.0

# Dev / test
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2
